pyproject.toml
[tool.poetry]
name = "tkinter-poetry-starter"
version = "1.0.0"
description = "A starter template for Python GUI applications using Tkinter and Poetry"
authors = ["ysskrishna <sivasaikrishnassk@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
Pillow = "^10.2.0"
tomli = "^2.0.1"
python-i18n = "^0.3.9"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
pylint = "^3.0.0"
pre-commit = "^3.6.0"
pyinstaller = "^6.4.0"

[tool.poetry.scripts]
tkinter_poetry_starter = "src.gui.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pylint.messages_control]
disable = ["C0111", "C0103"]

[tool.pylint.format]
max-line-length = 100
